outputs
This task demonstrates the concept of event handling in JavaScript through a simple interactive webpage. A button is created using HTML, and JavaScript is used to add functionality so that when the button is clicked, the background color of the entire webpage changes to a random color.

The core of this functionality is the use of a click event handler and a function that generates a random hexadecimal color. This helps beginners understand how JavaScript can be used to make webpages dynamic and responsive to user interactions.